Adjust views for nested layouts
I realized I wanted some layouts to live within a bootstrap
container,
but I wanted the static pages to live without one.
This took me a while to figure this out, but here's the section
of Rails Guides that finally explained it to me in a way that
I could understand:
https://guides.rubyonrails.org/layouts_and_rendering.html#using-nested-layouts

Within the app/views/layouts folder,
any html.erb file that has the same name as a controller
can serve as the layout for all pages within that controller.

Pretty neat!

I also added subtitles to all my views.