Reorganize views to extract active nav link script
All authenticated pages should now be nested
within manage_orders.html.erb, and they should
have the active nav link of "Manage Orders".

A bonus of this change is that, for the Static Pages,
the script now renders truly at the bottom of the page body,
instead of at the bottom of the content section.